+++ src/service/types/mod.rs
@@ -22,6 +22,8 @@ pub use list_versions_input::ListVersionsInput;
 pub use paginated_result::PaginatedResult;
 pub use persist_options::{PersistOptions, PersistOptionsBuilder};
 pub use search_documents_input::SearchDocumentsInput;
-pub use service_context::{Def, EventQueue, ServiceContext, ServiceContextBuilder, flush_queue};
+pub use service_context::{
+    Def, EmailContext, EventQueue, ServiceContext, ServiceContextBuilder, flush_queue,
+};
 pub use write_input::{WriteInput, WriteInputBuilder};
 pub use write_result::WriteResult;